Fix building POPT directly from CVS sources.
The ultimate golden rule of "the content of versioned files should be
_never_ changed during a regular building from source" was horribly
broken by gettextize which assumes that if po/Makefile.in.in is still
not present (which is the case after a checkout as we do not version
such a generated file) it can feel free to update Makefile.am and
configure.am, even if it doubles entries there.

Here skipping the call to gettextize in case po/Makefile.in.in does
not exists doesn't really help _initially_ at all. It just helps on
subsequent calls. To no longer break the rule one either has to fix
gettextize upstream (it should not blindly append po/Makefile.in.in to
"AC_OUTPUT" in configure.ac and "po" to "SUBDIRS" in Makefile.am in
if it is _already present_ there) or we have to workaround locally by
explicitly reversing its results. As we do not control gettextize, we
now reverse the results. I have chosen to use perl(1) instead of sed(1)
here as perl(1) is already used by Automake (so it is available anyway)
and it allows us to revert the changes with shorter calls on the CLI.

Additionally, let's call gettextize _after_ libtoolize to prevent
gettextize from complaining about a missing config.guess and config.sub.
Also, add config.rpath to Makefile.am as this is a required file,
generated by gettextize, too. Finally, pass --no-changelog to gettextize
as we really don't need a useless generated m4/ChangeLog distributed
with POPT.

This way one can now (again) do a fresh checkout of the popt/ module
and build POPT really out-of-the-box and without having versioned files
garbled up, too.