Fix iconv(3) usage.
POPT assumed that iconv(3) is available all the time. This is not
the case. It _might_ be available (and then indicated by the #define
HAVE_ICONV from gettext) if NLS is enabled. But if NLS is disabled
(--disable-nls) then the gettext NLS stuff not even _checks_ for an
available iconv(3) and hence it cannot even be used with an explicit
--with-libiconv-prefix, etc. Hence we POPT can use iconv(3) only for its
UTF-8 fiddling if HAVE_ICONV is defined.